addressed issue #3310: moved GATKProtectedMathUtils to MathUtils
