+++ src/main/java/org/broadinstitute/hellbender/tools/walkers/annotator/FragmentLength.java
@@ -3,7 +3,7 @@ package org.broadinstitute.hellbender.tools.walkers.annotator;
 import com.google.common.primitives.Ints;
 import htsjdk.variant.variantcontext.VariantContext;
 import org.broadinstitute.barclay.help.DocumentedFeature;
-import org.broadinstitute.hellbender.utils.GATKProtectedMathUtils;
+import org.broadinstitute.hellbender.utils.MathUtils;
 import org.broadinstitute.hellbender.utils.Utils;
 import org.broadinstitute.hellbender.utils.help.HelpConstants;
 import org.broadinstitute.hellbender.utils.read.GATKRead;
@@ -24,7 +24,7 @@ public class FragmentLength extends PerAlleleAnnotation implements StandardMutec
 
     @Override
     protected int aggregate(final List<Integer> values) {
-        return values.isEmpty() ? 0 : GATKProtectedMathUtils.median(Ints.toArray(values));
+        return values.isEmpty() ? 0 : MathUtils.median(Ints.toArray(values));
     }
 
     @Override
