+++ src/main/java/org/broadinstitute/hellbender/tools/walkers/mutect/SomaticLikelihoodsEngine.java
@@ -1,16 +1,12 @@
 package org.broadinstitute.hellbender.tools.walkers.mutect;
 
 import com.google.common.annotations.VisibleForTesting;
-import htsjdk.variant.variantcontext.Allele;
 import org.apache.commons.math3.linear.RealMatrix;
 import org.apache.commons.math3.special.Gamma;
 import org.apache.commons.math3.util.MathArrays;
 import org.broadinstitute.hellbender.utils.*;
-import org.broadinstitute.hellbender.utils.genotyper.LikelihoodMatrix;
-import org.broadinstitute.hellbender.utils.genotyper.ReadLikelihoods;
 
 import java.util.Arrays;
-import java.util.List;
 
 /**
  * Created by David Benjamin on 3/9/17.
@@ -60,13 +56,13 @@ public class SomaticLikelihoodsEngine {
     @VisibleForTesting
     protected static double[] getEffectiveCounts(RealMatrix log10Likelihoods, double[] dirichletPrior) {
         final double[] effectiveLog10Weights = new Dirichlet(dirichletPrior).effectiveLog10MultinomialWeights();
-        return GATKProtectedMathUtils.sumArrayFunction(0, log10Likelihoods.getColumnDimension(),
-                read -> GATKProtectedMathUtils.posteriors(effectiveLog10Weights, log10Likelihoods.getColumn(read)));
+        return MathUtils.sumArrayFunction(0, log10Likelihoods.getColumnDimension(),
+                read -> MathUtils.posteriors(effectiveLog10Weights, log10Likelihoods.getColumn(read)));
     }
 
     // same but with flat prior
     public static double[] getEffectiveCounts(RealMatrix log10Likelihoods) {
-        return GATKProtectedMathUtils.sumArrayFunction(0, log10Likelihoods.getColumnDimension(),
+        return MathUtils.sumArrayFunction(0, log10Likelihoods.getColumnDimension(),
                 read -> MathUtils.normalizeFromLog10ToLinearSpace(log10Likelihoods.getColumn(read)));
     }
 
@@ -86,7 +82,7 @@ public class SomaticLikelihoodsEngine {
 
         final double likelihoodsAndEntropyContribution = new IndexRange(0, log10Likelihoods.getColumnDimension()).sum(r -> {
             final double[] log10LikelihoodsForRead = log10Likelihoods.getColumn(r);
-            final double[] responsibilities = GATKProtectedMathUtils.posteriors(log10AlleleFractions, log10LikelihoodsForRead);
+            final double[] responsibilities = MathUtils.posteriors(log10AlleleFractions, log10LikelihoodsForRead);
             final double likelihoodsContribution = MathUtils.sum(MathArrays.ebeMultiply(log10LikelihoodsForRead, responsibilities));
             final double entropyContribution = Arrays.stream(responsibilities).map(SomaticLikelihoodsEngine::xLog10x).sum();
             return likelihoodsContribution - entropyContribution;
