+++ src/test/java/org/broadinstitute/hellbender/tools/copynumber/denoising/GCBiasCorrectorUnitTest.java
@@ -6,7 +6,7 @@ import org.apache.commons.math3.linear.Array2DRowRealMatrix;
 import org.apache.commons.math3.linear.DefaultRealMatrixChangingVisitor;
 import org.apache.commons.math3.linear.RealMatrix;
 import org.broadinstitute.hellbender.GATKBaseTest;
-import org.broadinstitute.hellbender.utils.GATKProtectedMathUtils;
+import org.broadinstitute.hellbender.utils.MathUtils;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 
@@ -61,7 +61,7 @@ public final class GCBiasCorrectorUnitTest extends GATKBaseTest {
         //
         final RealMatrix correctedCoverage = readCounts.copy();
         GCBiasCorrector.correctGCBias(correctedCoverage, intervalGCContent);
-        final double[] correctedNoiseBySample = GATKProtectedMathUtils.rowStdDevs(correctedCoverage);
+        final double[] correctedNoiseBySample = MathUtils.rowStdDevs(correctedCoverage);
         Arrays.stream(correctedNoiseBySample).forEach(x -> Assert.assertTrue(x < NON_GC_BIAS_NOISE_LEVEL * MEAN_READ_DEPTH));
 
         //check that GC-bias correction is approximately idempotent -- if you correct again, very little should happen
